Refactor image helper
